In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	int num1, num2, nextnum, i;
+	long num1, num2, nextnum, i;
 
 	num1 = 1;
 	num2 = 2;
